requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0